+++ tests/testthat/test-splines.R
@@ -115,11 +115,11 @@ test_that("CRAN basic spline tests", {
                     g = sample(state.abb[1:5], 100, replace = T),
                     f = sample(letters[1:5], 100, replace = T))
   
-  dat$y <- rbinom(100, 1, plogis(dat$x * runif(5)[match(dat$f, letters)]))
+  dat$y <- rbinom(100, 1, plogis(sin(sqrt(dat$x2 * 4)) + cos(dat$x * 3) * runif(5)[match(dat$f, letters)]))
   
   
   m1 <- vglmer(y ~ x + x2 + v_s(x), 
-               control = vglmer_control(iteration = 5),
+               control = vglmer_control(iteration = NITER),
                data = dat, family = 'binomial')
   expect_gt(min(diff(ELBO(m1, 'traj'))), - sqrt(.Machine$double.eps))  
   
@@ -127,12 +127,12 @@ test_that("CRAN basic spline tests", {
   # Check runs with 2
   m2 <- vglmer(y ~ v_s(x2) + v_s(x), 
                control = vglmer_control(
-                 iterations = NITER, print_prog = 20, prior_variance = 'mean_exists'),
+                 iterations = NITER, print_prog = 20),
                data = dat, family = 'binomial')
   # Check runs with "by"
   m3 <- vglmer(y ~ v_s(x2) + v_s(x, by = f), data = dat, 
                family = 'binomial',
-               control = vglmer_control(iterations = NITER, print_prog = 20))
+               control = vglmer_control(iterations = NITER, print_prog = 20, prior_variance = 'mean_exists'))
   # Check runs with RE 
   m4 <- vglmer(y ~ v_s(x, by = f) + (1 | g), 
                data = dat, family = 'binomial',
@@ -154,6 +154,15 @@ test_that("CRAN basic spline tests", {
      control = vglmer_control(iterations = NITER, print_prog = 20,
       factorization_method = 'strong'))
 
+  m8 <- vglmer(y ~ v_s(x2, by = g) + v_s(x, by = f, by_re = FALSE), 
+               control = vglmer_control(
+                 iterations = NITER, print_prog = 20),
+               data = dat, family = 'binomial')
+  
+  # Check that they are all accepted
+  expect_equal(m1$ELBO$accepted_PX, 1)
+  expect_equal(m2$ELBO$accepted_PX, 1)
+  
   # Check ELBO increases
   expect_gt(min(diff(m1$ELBO_trajectory$ELBO)), -sqrt(.Machine$double.eps))
   expect_gt(min(diff(m2$ELBO_trajectory$ELBO)), -sqrt(.Machine$double.eps))
@@ -162,15 +171,21 @@ test_that("CRAN basic spline tests", {
   expect_gt(min(diff(m5$ELBO_trajectory$ELBO)), -sqrt(.Machine$double.eps))
   expect_gt(min(diff(m6$ELBO_trajectory$ELBO)), -sqrt(.Machine$double.eps))
   expect_gt(min(diff(m7$ELBO_trajectory$ELBO)), -sqrt(.Machine$double.eps))
+  expect_gt(min(diff(m8$ELBO_trajectory$ELBO)), -sqrt(.Machine$double.eps))
   
-  expect_vector(predict(m1, newdata = dat[1:5,]))
-  expect_vector(predict(m2, newdata = dat[1:5,]))
-  expect_vector(predict(m3, newdata = dat[1:5,]))
-  expect_vector(predict(m4, newdata = dat[1:5,]))
-  expect_vector(predict(m5, newdata = dat[1:5,]))
-  expect_vector(predict(m6, newdata = dat[1:5,]))
-  expect_vector(predict(m7, newdata = dat[1:5,]))
-  
+  for (v in c(TRUE, FALSE)){
+    dat_custom <- data.frame(x = mean(dat$x), x2 = mean(dat$x2), 
+       g = c('CA', 'AL'), f = c('e', 'a'),
+       stringsAsFactors = v)
+    expect_vector(predict(m1, newdata = dat_custom))
+    expect_vector(predict(m2, newdata = dat_custom))
+    expect_vector(predict(m3, newdata = dat_custom))
+    expect_vector(predict(m4, newdata = dat_custom))
+    expect_vector(predict(m5, newdata = dat_custom))
+    expect_vector(predict(m6, newdata = dat_custom))
+    expect_vector(predict(m7, newdata = dat_custom))
+    expect_vector(predict(m8, newdata = dat_custom))
+  }
 })
 
 test_that("Test order of splines", {
@@ -186,7 +201,9 @@ test_that("Test order of splines", {
   m1 <- vglmer(y ~ x + x2 + v_s(x), 
      data = dat, family = 'binomial',
      control = vglmer_control(iterations = NITER))
+  
   expect_equal(length(coef(m1)), 3)
+  
   m1a <- vglmer(y ~ x2 + x + v_s(x), data = dat, 
     family = 'binomial', control = vglmer_control(iterations = NITER))
   
@@ -206,17 +223,97 @@ test_that("Prediction spline test", {
                     g = sample(state.abb[1:5], 100, replace = T),
                     f = sample(letters[1:5], 100, replace = T))
   
-  dat$y <- rbinom(100, 1, plogis(dat$x * runif(5)[match(dat$f, letters)]))
+  dat$y <- rbinom(100, 1, plogis(sin(3 * dat$x) * rnorm(5)[match(dat$f, letters)]))
   
-  m1 <- vglmer(y ~ x + x2 + v_s(x), 
-               data = dat, family = 'linear',
-               control = vglmer_control(iterations = 3))
-  # Check that prediction works for simple spline case
-  expect_vector(predict(m1, newdata = dat))
+  if (env_test != 'CRAN'){
+    loop_vector <- c('tpf', 'o')
+  }else{
+    loop_vector <- 'tpf'
+  }
   
-  
+  for (loop_type in loop_vector){
+    message(loop_type)
+    m1 <- vglmer(y ~ x2 + v_s(x, type = loop_type), 
+                 data = dat, family = 'linear',
+                 control = vglmer_control(iterations = NITER))
+    
+    # Check that prediction works for simple spline case
+    predict_dat <- expand.grid(x = seq(-3, 3, length.out=100), x2 = 0)
+    predict_vglmer <- predict(m1, newdata = predict_dat)
+    # Check that *fitted* aligns with *predict*  
+    expect_equivalent(fitted(m1), predict(m1, newdata = dat))
+    # Check that there is some slope in the function
+    diff_in_diff <- range(diff(diff(predict_vglmer)))
+    expect_gte(max(abs(diff_in_diff)), sqrt(.Machine$double.eps))
+    
+    raw_X <- vglmer_build_spline(x = predict_dat$x, knots = m1$internal_parameters$spline$attr[[1]]$knots,
+                                 type = m1$internal_parameters$spline$attr[[1]]$type,
+                                 Boundary.knots = m1$internal_parameters$spline$attr[[1]]$Boundary.knots, by = NULL)[[1]]$x
+    direct_predict <- as.vector(raw_X %*% m1$alpha$mean + cbind(1, 0, predict_dat$x) %*% m1$beta$mean)
+    expect_equivalent(direct_predict, predict_vglmer)    
+  }
+
+  if (TRUE){
+    
+    for (loop_type in loop_vector){
+      message(loop_type)
+      m1 <- vglmer(y ~ x2 + v_s(x, by = f, type = loop_type), 
+                   data = dat, family = 'linear',
+                   control = vglmer_control(iterations = NITER))
+      
+      # Check that prediction works for simple spline case
+      predict_dat <- expand.grid(x = seq(-3, 3, length.out=100), x2 = 0, f = c('a', 'd', 'e', 'c'))
+      predict_vglmer <- predict(m1, newdata = predict_dat)
+      # Check that *fitted* aligns with *predict*  
+      expect_equivalent(fitted(m1), predict(m1, newdata = dat))
+      # Check that there is some slope in the function
+      diff_in_diff <- sapply(split(predict_vglmer, predict_dat$f), FUN=function(i){
+        diff_in_diff <- diff(diff(i))/diff(range(i))
+        return(max(abs(diff_in_diff)))        
+      })
+      expect_true(all(diff_in_diff > sqrt(.Machine$double.eps)))
+      
+      raw_X <- vglmer_build_spline(x = predict_dat$x, knots = m1$internal_parameters$spline$attr[[1]]$knots,
+       type = m1$internal_parameters$spline$attr[[1]]$type,
+       by = predict_dat$f,
+       Boundary.knots = m1$internal_parameters$spline$attr[[1]]$Boundary.knots)[[1]]$x
+      
+      expand_x <- do.call('cbind', lapply(c('a', 'c', 'd', 'e'), FUN=function(i){
+        drop0(Diagonal(x = predict_dat$f == i) %*% raw_X)
+      }))
+      all_X <- cbind(raw_X, expand_x)
+      
+      direct_predict_spline <- as.vector(all_X %*% m1$alpha$mean[grep(rownames(m1$alpha$mean), pattern='spline @ x @ (base|[adec])'),] + 
+                cbind(1, 0, predict_dat$x) %*% m1$beta$mean)
+      
+      wide_alpha <- matrix(m1$alpha$mean[grepl(rownames(m1$alpha$mean), pattern='^f'),,drop=F], byrow = TRUE, ncol = 2)
+      
+      direct_predict <- rowSums(wide_alpha[match(predict_dat$f, c('a', 'b', 'c', 'd', 'e')),] * cbind(1, predict_dat$x))
+      expect_equivalent(direct_predict + direct_predict_spline, predict_vglmer)    
+    }
+  }
+
 })
 
+test_that("Compare against mgcv", {
+  
+  if (env_test == 'local'){
+    N <- 1000
+    dat <- data.frame(x = rnorm(N), x2 = rexp(N),
+                      g = sample(state.abb[1:5], N, replace = T),
+                      f = sample(letters[1:5], N, replace = T))
+    
+    dat$y <- rbinom(N, 1, plogis(2 * sin(3 * dat$x) + rnorm(5)[match(dat$f, letters)]))
+    
+    est_gam <- mgcv::gam(y ~ s(x) + s(f, bs = 're'), data = dat %>% mutate(f = factor(f)), family = binomial())  
+    est_vglmer <- vglmer(y ~ v_s(x, type = 'o') + (1 | f), data = dat, family = 'binomial')
+    
+    pred_estgam <- predict(est_gam, newdata = data.frame(x = seq(-3, 3, length.out=100), f = 'b'))
+    pred_estvglmer <- predict(est_vglmer, newdata = data.frame(x = seq(-3, 3, length.out=100), f = 'b'))
+    
+    expect_gte(cor(pred_estvglmer, pred_estgam), 0.95)
+  }
+})
 # TO-DO tests
 # check decomposition of D and then re-transformation
 # test with custom knots and prediction outside of knots/etc.