+++ NEWS.md
@@ -1,3 +1,8 @@
+# vglmer 1.0.2
+
+* IMPORTANT: Bug fix where prediction with only one spline was incorrect; the non-linear part was ignored.
+* Smaller bug fixes around splines have been added as well as updated tests.
+
 # vglmer 1.0.1
 
 * Patch to address compiler issues on CRAN